requirements.txt
pyrogram==2.0.106
tgcrypto
motor
feedparser
python-dotenv
apscheduler
aiohttp